TMS_Project/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
config==0.5.0.post0
Django==3.1.7
django-adminlte3==0.1.6
idna==2.10
psycopg2-binary==2.8.6
pytz==2021.1
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
